Create BaseModel
